RTC_ID:975329
COMMENT:[Map]同步所有代码

PatchModule:Navi

+++ app/src/main/java/com/fy/navi/hmi/mapdata/near/NearMapDataFragment.java
@@ -51,19 +51,31 @@ public class NearMapDataFragment extends BaseFragment<FragmentNearMapDataBinding
             @Override
             public void startAllTask(final ArrayList<Integer> cityAdCodes) {
                 Logger.d( "onItemClick cityAdCodes = " + GsonUtils.toJson(cityAdCodes));
-                mViewModel.startAllTask(cityAdCodes);
+                ThreadManager.getInstance().postDelay(() -> {
+                    if (mViewModel != null) {
+                        mViewModel.startAllTask(cityAdCodes);
+                    }
+                }, 0);
             }
 
             @Override
             public void pauseAllTask(final ArrayList<Integer> cityAdCodes) {
                 Logger.d( "pauseAllTask cityAdCodes = " + GsonUtils.toJson(cityAdCodes));
-                mViewModel.pauseAllTask(cityAdCodes);
+                ThreadManager.getInstance().postDelay(() -> {
+                    if (mViewModel != null) {
+                        mViewModel.pauseAllTask(cityAdCodes);
+                    }
+                }, 0);
             }
 
             @Override
             public void deleteAllTask(final ArrayList<Integer> cityAdCodes) {
                 Logger.d( "deleteAllTask cityAdCodes = " + GsonUtils.toJson(cityAdCodes));
-                mViewModel.deleteAllTask(cityAdCodes);
+                ThreadManager.getInstance().postDelay(() -> {
+                    if (mViewModel != null) {
+                        mViewModel.deleteAllTask(cityAdCodes);
+                    }
+                }, 0);
             }
 
         });
